# server/Startup/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Messages/project/{projectId}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "operationId": "Messages_GetProjectMessages",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/projects/{projectId}/photos": {
      "get": {
        "tags": [
          "AdminProjectPhotos"
        ],
        "operationId": "AdminProjectPhotos_GetPhotos",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "milestoneId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 3
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationItemsResponseOfPhotoDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminProjectPhotos"
        ],
        "operationId": "AdminProjectPhotos_UploadPhoto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  },
                  "MilestoneId": {
                    "type": "string",
                    "format": "guid",
                    "nullable": true
                  },
                  "Caption": {
                    "type": "string",
                    "nullable": true
                  },
                  "TakenAt": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/projects/{projectId}/photos/{photoId}": {
      "put": {
        "tags": [
          "AdminProjectPhotos"
        ],
        "operationId": "AdminProjectPhotos_UpdatePhoto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "photoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhotoDto"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminProjectPhotos"
        ],
        "operationId": "AdminProjectPhotos_DeletePhoto",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "photoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "projectId": {
            "type": "string",
            "format": "guid"
          },
          "senderId": {
            "type": "string",
            "format": "guid"
          },
          "senderName": {
            "type": "string"
          },
          "senderRole": {
            "type": "string"
          },
          "receiverId": {
            "type": "string",
            "format": "guid"
          },
          "content": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "attachmentUrl": {
            "type": "string",
            "nullable": true
          },
          "attachmentType": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "PaginationItemsResponseOfPhotoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PhotoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "projectId": {
            "type": "string",
            "format": "guid"
          },
          "milestoneId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "milestoneTitle": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "height": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "nullable": true
          },
          "takenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "uploadedBy": {
            "type": "string",
            "format": "guid"
          },
          "uploadedByName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "UpdatePhotoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "milestoneId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "caption": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "takenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}